Added the material compiler utility
Created the `compile` utility, which follows an input script to generate
a replacement texture pack that JkGfxMod will inject into the game.
Significant refactoring must be done in `renderer` to complete material
injection, in addition to some more shader variants (possibly an
ubershader). This work will be done in a future commit.