Additional compile diagnostics; fixed crash showing credits
Now, compile prints an error if any textures are non-pow2. This provides
some additional early validation for material packs.

Added some missing functions that were causing the game to crash when
displaying credits. This is not fully implemented yet; end credits show
as a black screen, but at least do not crash the game anymore.