Added internal screen resolution as config option
This configuration option allows JkGfxMod to trick the game into
thinking it is rendering at a different (internal) resolution than the
native screen resolution. Triangle rasterization is still performed at
the native screen resolution. The game display fills the screen, and is
letterboxed or pillarboxed to preserve the original aspect ratio.

Example:

JK doesn't work well with aspect ratios greater than 4:3. In order to
play the game in 4:3 aspect ratio on a 4K monitor, you need to set the
resolution to 2880x2160. Monitors generally don't support this
resolution.

With this change, you can set the resolution configuration option to
3840x2160, which is the native 4K display resolution. Then, you set the
internal resolution configuration option to 2880x2160, which is the
closest 4:3 resolution. JK will think your monitor is running at
2880x2160, but the screen will actually be set to 3840x2160, and the
game will be pillarboxed.