Suppressed out-of-band mode changes patched into the Steam version
The Steam version was modified to periodically call
ChangeDisplaySettingsA inside menu rendering functions. Ostensibly this
is to force the game to show the menu in fullscreen on modern computers,
but since JkGfxMod already handles this it is more annoying than
anything.

The call is just being suppressed for now.

This change also adds a configuration option, "log_path", which is
useful for turning on logging when it is inconvenient to change
environment variables.

Comment #11