Fixed some filesystem search bugs in compile
Two semi-related issues:
- Previously, compile couldn't locate a colormap file in a subdirectory.
This was due to mistaken use of std::directory_iterator instead of
std::recursive_directory_iteratory in some code recycled from Gorc.
- Compile also couldn't locate a colormap file on the filesystem if it
had a non-lowercased extension. This was due to another mistaken
assumption about this recycled code, that it automatically lowercased
all filenames in the VFS (true for GOBs, not true for filesystem). The
scan has been replaced with a case-insensitive alternative.